requirements.txt
adbutils==0.4.0
airtest==1.1.3
apkutils2==1.0.0
appdirs==1.4.3
asgiref==3.2.7
attrdict==2.0.1
attrs==19.1.0
beautifulsoup4==4.9.0
bs4==0.0.1
cached-property==1.5.1
certifi==2018.4.16
chardet==3.0.4
cigam==0.0.3
colorama==0.4.1
comtypes==1.1.7
cssselect==1.1.0
cycler==0.10.0
ddt==1.2.1
decorator==4.4.0
Deprecated==1.2.10
deprecation==2.0.6
diff-match-patch==20121119
Django==2.0.5
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.0.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.13
django-simple-captcha==0.5.9
djangorestframework==3.8.2
DjangoUeditor==1.8.143
docopt==0.6.2
et-xmlfile==1.0.1
facebook-wda==0.3.4
fake-useragent==0.1.11
fire==0.2.0
future==0.16.0
get==1.0.3
httplib2==0.9.2
humanize==0.5.1
idna==2.6
imageio==2.8.0
jdcal==1.4
Jinja2==2.11.2
kiwisolver==1.1.0
logzero==1.5.0
Markdown==2.6.11
MarkupSafe==1.1.1
matplotlib==3.1.1
mss==4.0.3
networkx==2.4
numpy==1.16.4
odfpy==1.3.6
olefile==0.46
opencv-contrib-python==3.4.2.17
opencv-python==4.2.0.34
openpyxl==2.5.3
packaging==19.1
parse==1.15.0
Pillow==4.0.0
post==1.0.2
progress==1.5
public==1.0.3
py==1.8.0
pycryptodomex==3.8.2
pyee==7.0.1
pyelftools==0.26
PyMySQL==0.8.1
pyparsing==2.4.0
pypiwin32==223
pyppeteer==0.0.25
pyquery==1.4.1
pytesseract==0.2.6
python-dateutil==2.8.0
pytz==2018.4
PyWavelets==1.1.1
pywin32==224
pywinauto==0.6.3
PyYAML==3.12
query-string==1.0.2
request==1.0.2
requests==2.18.4
requests-html==0.10.0
retry==0.9.2
scikit-image==0.17.2
scipy==1.4.1
selenium==3.141.0
six==1.11.0
soupsieve==2.0
sqlparse==0.3.0
tablib==0.12.1
termcolor==1.1.0
tifffile==2020.6.3
tqdm==4.45.0
uiautomator2==2.9.5
unicodecsv==0.14.1
urllib3==1.22
w3lib==1.21.0
websockets==8.1
weditor==0.5.0
whichcraft==0.6.0
win32core==221.36
win32gui==221.6
wrapt==1.12.1
xadmin==2.0.1
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
